+++ src/main/java/org/team4/hanzip/api/member/dto/mypage/MyPageRequestDTO.java
@@ -1,0 +1,10 @@
+package org.team4.hanzip.api.member.dto.mypage;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class MyPageRequestDTO {
+    private Long memberId;
+}
