+++ src/main/java/org/team4/hanzip/api/member/dto/mypage/MyPageResponseDTO.java
@@ -1,0 +1,10 @@
+package org.team4.hanzip.api.member.dto.mypage;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@AllArgsConstructor
+@Getter
+public class MyPageResponseDTO {
+    private String nickname;
+}
