+++ src/main/java/org/team4/hanzip/api/member/dto/signup/SignUpResponseDTO.java
@@ -1,4 +1,4 @@
-package org.team4.hanzip.api.member.dto;
+package org.team4.hanzip.api.member.dto.signup;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
